Feature flags: Support "auto-enable" feature flags
A feature flag can be marked as "auto-enable" by setting `auto_enable` to
true in its properties.

An auto-enable feature flag is automatically enabled as soon as all
nodes in the cluster support it. This is achieved by trying to enable it
when RabbitMQ starts, when a plugin is enabled/disabled or when a node
joins/re-joins a cluster. If the feature flag can't be enabled, the
error is ignored.

An auto-enable feature flag also implicitly depends on
`feature_flags_v2`.

However, it should be used cautiously, especially if the feature flag
has a migration function, because it might be enabled at an
inappropriate time w.r.t. the user's workload.