Only look for tracked channels of a closed connection locally
`get_tracked_channels_by_connection_pid/1` is only called for a local
connection, so no need to RPC to all nodes in the cluster as the
channels should be on the same node.

This commit leaves the old behaviour unchanged and only affects code
path when feature flag `tracking_records_in_ets` is enabled.